# docs/04-api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Intent Rebase Engine API
  description: |
    Manages intent lifecycle, versioning, diff computation, rebase preview, and dependency graph operations.

    ## Implemented scopes

    ### Phase 1 — Core Control Plane
    - Health & observability endpoints
    - Intent lifecycle (create/read/version)
    - Intent analysis (validate, diff, rebase-preview)
    - Dependency graph CRUD

    ### Phase 3 Batch 1 — Compensation + Orchestration
    - Side effect capture and query
    - Compensation action CRUD and approve/waive/execute
    - DLQ and batch candidate APIs
    - Policy gate evaluation
    - Orchestration dashboard and coordination view
    - Orchestration dry-run planner and batch approve/reapprove/execute
    - Single-shot orchestration runtime (HTTP + CLI)
    - Compensation simulation: GET /intents/{intent_id}/rebase-simulation and POST /compensation-simulation/run (N4-4 bounded slice; commit fe2a1f6)

    ### Phase 3 Batch 3b — Forensic Replay Verification (bounded slice)
    - Request-driven forensic bundle feasibility verification with real entity counts (commit 7b05c5b)
    - Forensic bundle generation with S3/MinIO persistence seam (POST /forensic/bundle)
    - Forensic archive export (POST /forensic/export)
    - Forensic bundle list and download (GET /forensic/bundles, GET /forensic/bundles/{id}/download)
    - Read-only verification and integrity checking (no full runtime replay)

...

    ## Bounded scope note
...

...

...

...

...

...
  version: 0.1.0
...
servers:
...
paths:
  /intents:
    post:
      tags:
        - intents
      summary: Create a new intent
      description: Creates a new intent with initial version (v1)
      operationId: createIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntentRequest'
      responses:
        '201':
          description: Intent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIntentResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /intents/{intent_id}:
    get:
      tags:
        - intents
      summary: Get intent head (current version)
      description: Returns the intent document with its current (head) version
      operationId: getIntentHead
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Intent ID
      responses:
        '200':
          description: Intent found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentHeadResponse'
        '404':
          description: Intent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /intents/{intent_id}/versions:
    post:
      tags:
        - versions
      summary: Create a new version
      description: |
        Creates a new version of an existing intent.

        **Optimistic Concurrency Control (OCC):**
        - `X-Expected-Version` (optional): The version number the client believes is current.
          If provided and doesn't match, returns 409 Conflict.
          If header is present but not a valid integer, returns 400 Bad Request.
        - `X-Expected-Row-Version` (optional): The row_version the client last observed.
          If provided and doesn't match, returns 409 Conflict.
          If header is present but not a valid integer, returns 400 Bad Request.

        If headers are omitted, the service uses the current server state (current_version and
        row_version) as the expected values. This means the OCC check still occurs, but against
        whatever is currently stored — allowing concurrent overwrites if multiple clients omit headers.
        Clients that need deterministic OCC protection should provide both headers.
      operationId: createVersion
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Intent ID
        - name: X-Expected-Version
          in: header
          required: false
          schema:
            type: integer
          description: Expected current version for OCC check
        - name: X-Expected-Row-Version
          in: header
          required: false
          schema:
            type: integer
          description: Expected row_version for OCC check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVersionRequest'
      responses:
        '201':
          description: Version created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateVersionResponse'
        '400':
          description: Invalid request - malformed header or request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Intent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Concurrency conflict - intent has been modified since last read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - versions
      summary: List all versions
      description: Returns all versions of an intent in descending version order
      operationId: listVersions
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Intent ID
      responses:
        '200':
          description: Versions found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListVersionsResponse'
        '404':
          description: Intent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /intents/{intent_id}/versions/{version_number}:
    get:
      tags:
        - versions
      summary: Get specific version
      description: Returns a specific version of an intent by version number
      operationId: getVersion
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Intent ID
        - name: version_number
          in: path
          required: true
          schema:
            type: integer
          description: Version number
      responses:
        '200':
          description: Version found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentVersion'
        '404':
          description: Bundle not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /intents/{intent_id}/diff:
    post:
      tags:
        - diff
      summary: Compute diff between two versions
      description: |
        Computes a semantic diff between two versions of an intent.

        The diff covers scope, constraints, acceptance_criteria, and authority sections.
        Returns structured diff output plus risk analysis (severity, confidence, manual-review).

        **Version ordering:**
        - `from_version` must be less than `to_version`
        - Both versions must belong to the same intent
        - Both versions must exist
      operationId: computeDiff
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Intent ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiffRequest'
      responses:
        '200':
          description: Diff computed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffResponse'
        '400':
          description: Invalid request - bad version ordering or versions from different intents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Intent or version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /compensation-actions/{action_id}/waive:
    post:
      tags:
        - compensation
      summary: Waive a pending compensation action
      description: |
        Phase 3 Batch 1 (bounded execution slice): Transitions a Pending compensation action
        to Waived status, marking it as intentionally skipped.

        **Transition rules:**
        - Only Pending actions can be waived
        - Uses optimistic locking via lock_version to prevent concurrent updates

        **Fails closed on illegal transitions:**
        - Returns 409 Conflict if action is not Pending
        - Returns 409 Conflict if lock_version doesn't match

        **This slice:** Waived actions are terminal. No reactivation path exists.
      operationId: waiveCompensationAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Compensation action ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaiveCompensationActionBody'
      responses:
        '200':
          description: Compensation action waived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationActionResponse'
        '404':
          description: Compensation action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Compensation action is not Pending or lock_version mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    CreateIntentRequest:
      type: object
      required: [workflow_id, source_refs, payload, created_by, tags]
      properties:
        workflow_id:
          type: string
          format: uuid
        source_refs:
          type: array
          items:
            $ref: '#/components/schemas/SourceRef'
        payload:
          $ref: '#/components/schemas/IntentPayload'
        created_by:
          $ref: '#/components/schemas/ActorRef'
        tags:
          type: array
          items:
            type: string
...